requirements.txt
fastapi>=0.100.0
uvicorn>=0.22.0
databases>=0.7.0
python-dotenv>=1.0.0
oracledb>=1.3.0 
asyncpg>=0.27.0  
httpx>=0.24.0   
pytest>=7.3.1
